rename: @/components/common/fund => @/components/fund
